generate for dummy core, also in test_all.py
a lot of fixes. All looks alright, ready for review

+++ software/cryptotvgen/cryptotvgen/prepare_libs.py
@@ -17,6 +17,7 @@ except ImportError:
 
 # TODO change `print`s to appropriate log functions
 
+
 lwc_candidates = {'hash': ['ace', 'ascon', 'drygascon', 'esch', 'gimli', 'knot', 'photonbeetle',
                            'saturnin', 'skinnyhash', 'subterranean', 'xoodyak'],
                   'aead': ['ace', 'ascon', 'comet', 'drygascon', 'elephant', 'estate', 'paefforkskinny', 'giftcofb', 'gimli',
@@ -26,16 +27,6 @@ lwc_candidates = {'hash': ['ace', 'ascon', 'drygascon', 'esch', 'gimli', 'knot',
 
 mkfile_name = 'lwc_cffi.mk'
 
-def build_variants(variants, candidates_dir):
-    for vname, vtype in variants:
-        # print(f'running make CRYPTO_VARIANT={vname} CRYPTO_TYPE={vtype} in {ctgen_candidates_dir}')
-        cp = subprocess.run(['make', '-f',  str(candidates_dir / mkfile_name), '-C', str(candidates_dir),
-                             f'CRYPTO_VARIANT={vname}', f'CRYPTO_TYPE={vtype}'], cwd=candidates_dir)
-        try:
-            cp.check_returncode()
-        except:
-            print(f'`make CRYPTO_VARIANT={vname} CRYPTO_TYPE={vtype}` failed (exit code: {cp.returncode})')
-            sys.exit(1)
 
 def ctgen_data_dir(sub_dir=None):
     data_dir = pathlib.Path.home() / '.cryptotvgen'
@@ -46,8 +37,27 @@ def ctgen_data_dir(sub_dir=None):
     return data_dir
 
 
-def build_supercop_libs(sc_version, libs='all'):
+def build_supercop_libs(sc_version, libs='all', candidates_dir=None, lib_path=None):
     ctgen_candidates_dir = ctgen_data_dir()
+    
+    #TODO
+    impl_src_dir = 'ref'
+    
+    def build_variants(variants, candidates_dir):
+      for vname, vtype in variants:
+          # print(f'running make CRYPTO_VARIANT={vname} CRYPTO_TYPE={vtype} in {ctgen_candidates_dir}')
+          cmd = ['make', '-f',  str(ctgen_candidates_dir / mkfile_name), '-C', str(candidates_dir),
+                 f'CRYPTO_VARIANT={vname}', f'CRYPTO_TYPE={vtype}', f'CANDIDATE_PATH={candidates_dir}',
+                 f'IMPL_SRC_DIR={impl_src_dir}']
+          if lib_path:
+              print(f"binaries will be in LIB_PATH={lib_path}")
+              cmd.append(f'LIB_PATH={lib_path}')
+          cp = subprocess.run(cmd, cwd=candidates_dir)
+          try:
+              cp.check_returncode()
+          except:
+              print(f'`{" ".join(cmd)}` failed! (exit code: {cp.returncode})')
+              sys.exit(1)
 
     if libs == 'all' or libs == ['all']:
         print('building all libs')
@@ -67,8 +77,9 @@ def build_supercop_libs(sc_version, libs='all'):
         tar_path = urllib.request.urlretrieve(sc_url, filename=tar_path)[0]
         print(f'Download successfull!')
         return tarfile.open(tar_path)
-
-    sc_tar = get_sc_tar(sc_version)
+    
+    def filter_variants(variants):
+      return variants # TODO
 
     ctgen_includes_dir = ctgen_data_dir('includes')
 
@@ -79,43 +90,64 @@ def build_supercop_libs(sc_version, libs='all'):
     mk_content = pkg_resources.read_text(__package__, mkfile_name)
     with open(ctgen_candidates_dir / mkfile_name, 'w') as f:
         f.write(mk_content)
-
-    incl_candidates = set()
-    extract_list = []
+    
     variants = set()
+    
+    if not candidates_dir:
+      sc_tar = get_sc_tar(sc_version)
 
-    crypto_dir_regexps = {crypto_type: re.compile(f'supercop-{sc_version}/crypto_{crypto_type}/([^/]+)/ref/')
-                          for crypto_type in lwc_candidates.keys()}
+      incl_candidates = set()
+      extract_list = []
 
-    # TODO make this more efficient, though unlikely to be a performance bottleneck
+      crypto_dir_regexps = {crypto_type: re.compile(f'supercop-{sc_version}/crypto_{crypto_type}/([^/]+)/{impl_src_dir}/')
+                            for crypto_type in lwc_candidates.keys()}
 
-    def match_tarinfo(tarinfo):
-        for crypto_type in lwc_candidates.keys():
-            match = crypto_dir_regexps[crypto_type].match(tarinfo.name)
-            if match:
-                variant_name = match.group(1)
-                for cnd in lwc_candidates[crypto_type]:
-                    if variant_name.startswith(cnd):
-                        variants.add((variant_name, crypto_type))
-                        extract_list.append(tarinfo)
-                        incl_candidates.add(cnd)
-                        return
+      # TODO make this more efficient, though unlikely to be a performance bottleneck
 
-    print('decompressing archive and determining the list of files to extract...')
-    for tarinfo in sc_tar:
-        match_tarinfo(tarinfo)
+      def match_tarinfo(tarinfo):
+          for crypto_type in lwc_candidates.keys():
+              match = crypto_dir_regexps[crypto_type].match(tarinfo.name)
+              if match:
+                  variant_name = match.group(1)
+                  for cnd in lwc_candidates[crypto_type]:
+                      if variant_name.startswith(cnd):
+                          variants.add((variant_name, crypto_type))
+                          extract_list.append(tarinfo)
+                          incl_candidates.add(cnd)
+                          return
 
-    candidates_not_found = set(lwc_candidates['aead'] + lwc_candidates['hash']) - incl_candidates
-    assert not candidates_not_found, f"The following candidates were not found in the SUPERCOP archive: {candidates_not_found}"
+      print('decompressing archive and determining the list of files to extract...')
+      for tarinfo in sc_tar:
+          match_tarinfo(tarinfo)
 
-    print('extracting files...')
-    tmp_dir = tempfile.mkdtemp()
-    sc_tar.extractall(path=tmp_dir, members=extract_list)
-    print('extraction complete')
+      candidates_not_found = set(lwc_candidates['aead'] + lwc_candidates['hash']) - incl_candidates
+      assert not candidates_not_found, f"The following candidates were not found in the SUPERCOP archive: {candidates_not_found}"
 
-    shutil.copytree(str(pathlib.Path(tmp_dir) / f'supercop-{sc_version}'),
-                    str(ctgen_candidates_dir), dirs_exist_ok=True)
-    shutil.rmtree(tmp_dir)
-    print('moved sources to cryptotvgen data dir')
+      print('extracting files...')
+      tmp_dir = tempfile.mkdtemp()
+      sc_tar.extractall(path=tmp_dir, members=extract_list)
+      print('extraction complete')
 
-    build_variants(variants, ctgen_candidates_dir)
+      shutil.copytree(str(pathlib.Path(tmp_dir) / f'supercop-{sc_version}'),
+                      str(ctgen_candidates_dir), dirs_exist_ok=True)
+      shutil.rmtree(tmp_dir)
+      print('moved sources to cryptotvgen data dir')
+      candidates_dir = ctgen_candidates_dir
+    else:
+      candidates_dir = pathlib.Path(candidates_dir)
+      for crypto_type in ['aead', 'hash']:
+        try:
+            dir_iter = (candidates_dir / f'crypto_{crypto_type}').iterdir()
+            for sub in dir_iter:
+                if sub.is_dir():
+                    impl = sub / impl_src_dir
+                    vname = sub.name
+                    if impl.exists() and impl.is_dir():
+                        print(f"found variant:{vname} ({crypto_type})")
+                        variants.add((vname, crypto_type))
+        except FileNotFoundError as e:
+            sys.exit(f"{e}\ncandidates_dir={candidates_dir} does not have a crypto_{crypto_type}/{impl_src_dir} sub directory!\n")
+            
+    variants = filter_variants(variants)
+
+    build_variants(variants, candidates_dir)
